backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]==0.41.0
python-dotenv==1.2.2
google-generativeai==0.8.6
google-cloud-storage>=3.0.0
openai>=1.0.0
requests>=2.31.0
gspread>=6.0.0

